Move tests to their respoective modules
