+++ src/libs/humanizer/humanizerModules.test.ts
@@ -1,4 +1,3 @@
-import { ethers } from 'ethers'
 import fetch from 'node-fetch'
 
 import { describe, expect, test } from '@jest/globals'
@@ -10,21 +9,12 @@ import { ErrorRef } from '../../controllers/eventEmitter/eventEmitter'
 import { AccountOp } from '../accountOp/accountOp'
 import { humanizeCalls, visualizationToText } from './humanizerFuncs'
 import { humanizerCallModules as humanizerModules } from './index'
-import { HumanizerFragment, HumanizerMeta, HumanizerVisualization, IrCall } from './interfaces'
-import { aaveHumanizer } from './modules/Aave'
-import { privilegeHumanizer } from './modules/privileges'
-import { sushiSwapModule } from './modules/sushiSwapModule'
-import { uniswapHumanizer } from './modules/Uniswap'
-// import { oneInchHumanizer } from './modules/oneInch'
-import { WALLETModule } from './modules/WALLET'
-import { wrappingModule } from './modules/wrapped'
+import { HumanizerFragment, HumanizerMeta, IrCall } from './interfaces'
 import { parseCalls } from './parsers'
 import { humanizerMetaParsing } from './parsers/humanizerMetaParsing'
-import { getAction, getLabel, getToken, integrateFragments } from './utils'
+import { integrateFragments } from './utils'
 
 const humanizerInfo = _humanizerInfo as HumanizerMeta
-const TETHER_ADDRESS = '0xdac17f958d2ee523a2206206994597c13d831ec7'
-const WETH_ADDRESS = '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2'
 const accountOp: AccountOp = {
   accountAddr: '0xB674F3fd5F43464dB0448a57529eAF37F04cceA5',
   networkId: 'ethereum',
@@ -279,446 +269,5 @@ describe('module tests', () => {
     const res = irCalls.map((call: IrCall) => visualizationToText(call, standartOptions))
     expect(expectedTexification.length).toBe(res.length)
     expectedTexification.forEach((et: string, i: number) => expect(res[i]).toEqual(et))
-  })
-  test('uniV3', () => {
-    const expectedhumanization = [
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0x8a3c710e41cd95799c535f22dbae371d7c858651',
-          amount: 50844919041919270406243n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-          amount: 137930462904193673n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0x0000000000000000000000000000000000000000'
-        },
-        { type: 'deadline', amount: 1692784103000n }
-      ],
-      [
-        { type: 'action', content: 'Unwrap' },
-        {
-          type: 'token',
-          address: '0x0000000000000000000000000000000000000000',
-          amount: 137930462904193673n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0xbc5a0707cc6c731debea1f0388a4240df93259e4'
-        }
-      ],
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-          amount: 941000000000000000n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0x6e975115250b05c828ecb8ededb091975fc20a5d',
-          amount: 5158707941840645403045n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0xbb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a8'
-        },
-        { type: 'deadline', amount: 1692786121000n }
-      ],
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0xebb82c932759b515b2efc1cfbb6bf2f6dbace404',
-          amount: 422775565331912310692n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
-          amount: 2454922038n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0xca124b356bf11dc153b886ecb4596b5cb9395c41'
-        },
-        { type: 'deadline', amount: 1692781751000n }
-      ],
-      [
-        { type: 'action', content: 'Swap up to' },
-        {
-          type: 'token',
-          address: '0x6e975115250b05c828ecb8ededb091975fc20a5d',
-          amount: 4825320403256397423633n
-        },
-        { type: 'label', content: 'for' },
-        {
-          type: 'token',
-          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-          amount: 941000000000000000n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0xbb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a8'
-        },
-        { type: 'deadline', amount: 1692784529000n }
-      ],
-      // problematic one
-      [
-        { type: 'action', content: 'Wrap' },
-        {
-          type: 'token',
-          address: '0x0000000000000000000000000000000000000000',
-          amount: 100000000000000n
-        }
-      ],
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0x4200000000000000000000000000000000000006',
-          amount: 100000000000000n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0x0b2c639c533813f4aa9d7837caf62653d097ff85',
-          amount: 178131n
-        },
-        { type: 'deadline', amount: 1699015475000n }
-      ],
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-          amount: 100000000n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0x2e9a6df78e42a30712c10a9dc4b1c8656f8f2879',
-          amount: 72003605256085551n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0x02a3109c4ce8354ee771feac419b5da04ef15761'
-        },
-        { type: 'deadline', amount: 1700232005000n }
-      ],
-      [
-        {
-          type: 'action',
-          content: 'Grant approval'
-        },
-        {
-          type: 'label',
-          content: 'for'
-        },
-        {
-          type: 'token',
-          address: '0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
-          amount: 1461501637330902918203684832716283019655932542975n
-        },
-        {
-          type: 'label',
-          content: 'to'
-        },
-
-        {
-          type: 'address',
-          address: '0x643770e279d5d0733f21d6dc03a8efbabf3255b4'
-        }
-      ],
-      [
-        { type: 'action', content: 'Swap' },
-        {
-          type: 'token',
-          address: '0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
-          amount: 10000n
-        },
-        { type: 'label', content: 'for at least' },
-        {
-          type: 'token',
-          address: '0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270',
-          amount: 10337979384133110n
-        },
-        { type: 'deadline', amount: 1708105881000n }
-      ],
-      [
-        { type: 'action', content: 'Unwrap' },
-        {
-          type: 'token',
-          address: '0x0000000000000000000000000000000000000000',
-          amount: 10337979384133110n
-        },
-        { type: 'label', content: 'and send it to' },
-        {
-          type: 'address',
-          address: '0x0000000000000000000000000000000000000001'
-        }
-      ]
-    ]
-    accountOp.calls = [...transactions.uniV3]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = uniswapHumanizer(accountOp, irCalls, humanizerInfo, {
-      emitedError: moockEmitError
-    })
-    expect(irCalls.length).toBe(expectedhumanization.length)
-    irCalls.forEach((c, i) => {
-      expect(c?.fullVisualization?.length).toEqual(expectedhumanization[i].length)
-      c?.fullVisualization?.forEach((v: HumanizerVisualization, j: number) => {
-        expect(v).toMatchObject(expectedhumanization[i][j])
-      })
-    })
-  })
-  test('WETH', () => {
-    accountOp.calls = [...transactions.weth]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = wrappingModule(accountOp, irCalls, humanizerInfo)
-    expect(irCalls[0].fullVisualization?.length).toBe(3)
-    expect(irCalls[0]?.fullVisualization![0]).toMatchObject({ type: 'action', content: 'Wrap' })
-    expect(irCalls[0]?.fullVisualization![1]).toMatchObject({
-      type: 'token',
-      address: '0x0000000000000000000000000000000000000000',
-      amount: transactions.weth[0].value
-    })
-    expect(irCalls[0]?.fullVisualization![2]).toMatchObject({
-      type: 'token',
-      address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      amount: 0n,
-      isHidden: true
-    })
-
-    expect(irCalls[1].fullVisualization?.length).toBe(3)
-    expect(irCalls[1]?.fullVisualization![0]).toMatchObject({ type: 'action', content: 'Unwrap' })
-    expect(irCalls[1]?.fullVisualization![1]).toMatchObject({
-      type: 'token',
-      address: '0x0000000000000000000000000000000000000000',
-      amount: 8900000000000000n
-    })
-    expect(irCalls[0]?.fullVisualization![2]).toMatchObject({
-      type: 'token',
-      address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-      amount: 0n,
-      isHidden: true
-    })
-  })
-
-  test('SWAP WRAP/UNWRAPS', () => {
-    const placeholder = '0x123456789'
-    const calls: IrCall[] = [
-      {
-        to: placeholder,
-        data: placeholder,
-        value: 0n,
-        fullVisualization: [
-          getAction('Swap'),
-          getToken(TETHER_ADDRESS, 1000000000n),
-          getLabel(placeholder),
-          getToken(WETH_ADDRESS, 1000000000n)
-        ]
-      },
-      {
-        to: WETH_ADDRESS,
-        data: placeholder,
-        value: 0n,
-        fullVisualization: [getAction('Unwrap'), getToken(ethers.ZeroAddress, 1000000000n)]
-      }
-    ]
-
-    const expectedHumanization = [
-      { type: 'action', content: 'Swap' },
-      {
-        type: 'token',
-        address: '0xdac17f958d2ee523a2206206994597c13d831ec7',
-        amount: 1000000000n
-      },
-      { type: 'label', content: '0x123456789' },
-      {
-        type: 'token',
-        address: '0x0000000000000000000000000000000000000000',
-        amount: 1000000000n
-      }
-    ]
-
-    const [newCalls] = wrappingModule(accountOp, calls, humanizerInfo)
-    expect(newCalls.length).toBe(1)
-    newCalls[0].fullVisualization?.map((v, i) => expect(v).toMatchObject(expectedHumanization[i]))
-  })
-
-  test('AAVE', () => {
-    const expectedhumanization = [
-      [
-        { content: 'Deposit' },
-        { type: 'token' },
-        { content: 'to Aave lending pool' },
-        { content: 'on behalf of' },
-        { type: 'address' }
-      ],
-      [
-        { content: 'Withdraw' },
-        { type: 'token' },
-        { content: 'from Aave lending pool' },
-        { content: 'on behalf of' },
-        { type: 'address' }
-      ],
-      [
-        { content: 'Deposit' },
-        { type: 'token' },
-        { content: 'to Aave lending pool' },
-        { content: 'on behalf of' },
-        { type: 'address' }
-      ],
-      [
-        { content: 'Withdraw' },
-        { type: 'token' },
-        { content: 'from Aave lending pool' },
-        { content: 'on behalf of' },
-        { type: 'address' }
-      ]
-    ]
-    accountOp.calls = [...transactions.aaveLendingPoolV2, ...transactions.aaveWethGatewayV2]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = aaveHumanizer(accountOp, irCalls, humanizerInfo)
-    irCalls.forEach((c, i) =>
-      c?.fullVisualization?.forEach((v: HumanizerVisualization, j: number) =>
-        expect(v).toMatchObject(expectedhumanization[i][j])
-      )
-    )
-  })
-  test('WALLET', () => {
-    const expectedhumanization = [
-      [
-        { type: 'action', content: 'Deposit' },
-        {
-          type: 'token',
-          address: '0x88800092ff476844f74dc2fc427974bbee2794ae',
-          amount: 10000000000000000000000n
-        },
-        { type: 'label', content: 'to' },
-        {
-          type: 'address',
-          address: '0x47cd7e91c3cbaaf266369fe8518345fc4fc12935'
-        }
-      ],
-      [
-        { type: 'action', content: 'Leave' },
-        { type: 'label', content: 'with' },
-        {
-          type: 'token',
-          address: '0x88800092ff476844f74dc2fc427974bbee2794ae',
-          amount: 2527275889852892335882193n
-        },
-        {
-          type: 'address',
-          address: '0x47cd7e91c3cbaaf266369fe8518345fc4fc12935'
-        }
-      ],
-      [
-        { type: 'action', content: 'Rage leave' },
-        { type: 'label', content: 'with' },
-        {
-          type: 'token',
-          address: '0x88800092ff476844f74dc2fc427974bbee2794ae',
-          amount: 2019750399052452828721n
-        },
-        {
-          type: 'address',
-          address: '0x47cd7e91c3cbaaf266369fe8518345fc4fc12935'
-        }
-      ]
-    ]
-    accountOp.calls = [...transactions.WALLET]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = WALLETModule(accountOp, irCalls, humanizerInfo)
-
-    irCalls.forEach((c, i) =>
-      c?.fullVisualization?.forEach((v: HumanizerVisualization, j: number) =>
-        expect(v).toMatchObject(expectedhumanization[i][j])
-      )
-    )
-  })
-
-  test('SushiSwap RouteProcessor', () => {
-    const expectedhumanization: Partial<HumanizerVisualization>[] = [
-      { type: 'action', content: 'Swap' },
-      {
-        type: 'token',
-        address: '0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270',
-        amount: 400000000000000n
-      },
-      { type: 'label', content: 'for' },
-      {
-        type: 'token',
-        address: '0x8f3cf7ad23cd3cadbd9735aff958023239c6a063',
-        amount: 348830169184669n
-      }
-    ]
-    accountOp.calls = [...transactions.sushiSwapCalls]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = sushiSwapModule(accountOp, irCalls, humanizerInfo)
-    expect(irCalls.length).toBe(1)
-    expectedhumanization.forEach((h: Partial<HumanizerVisualization>, i: number) => {
-      expect(irCalls[0]?.fullVisualization?.[i]).toMatchObject(h)
-    })
-  })
-
-  test('Privilege Humanizer', async () => {
-    const expectedhumanization: Partial<HumanizerVisualization>[][] = [
-      [
-        { type: 'action', content: 'Enable' },
-        {
-          type: 'address',
-          address: '0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789'
-        }
-      ],
-      [
-        { type: 'action', content: 'Update access status' },
-        { type: 'label', content: 'of' },
-        {
-          type: 'address',
-          address: '0x6969174fd72466430a46e18234d0b530c9fd5f49'
-        },
-        { type: 'label', content: 'to' },
-        {
-          type: 'label',
-          content: 'regular access'
-        }
-      ],
-      [
-        { type: 'action', content: 'Revoke access' },
-        { type: 'label', content: 'of' },
-        {
-          type: 'address',
-          address: '0x6969174fd72466430a46e18234d0b530c9fd5f49'
-        }
-      ]
-    ]
-    accountOp.calls = [...transactions.privileges]
-    let irCalls: IrCall[] = accountOp.calls
-    ;[irCalls] = privilegeHumanizer(accountOp, irCalls, humanizerInfo)
-
-    expect(irCalls.length).toBe(expectedhumanization.length)
-    expectedhumanization.forEach(
-      (callHumanization: Partial<HumanizerVisualization>[], i: number) => {
-        callHumanization.forEach((h: Partial<HumanizerVisualization>, j: number) =>
-          expect(irCalls[i]?.fullVisualization?.[j]).toMatchObject(h)
-        )
-      }
-    )
   })
 })
