+++ src/libs/humanizer/modules/GasTankModule/gasTank.test.ts
@@ -1,0 +1,9 @@
+import { describe } from '@jest/globals'
+
+// @TODO
+
+describe('gasTank', () => {
+  test('basic', () => {
+    // @TODO
+  })
+})
