+++ src/libs/humanizer/modules/GasTankModule/index.ts
@@ -1,0 +1,3 @@
+import { gasTankModule } from './gasTankModule'
+
+export default gasTankModule
