+++ src/libs/humanizer/modules/PreProcess/index.ts
@@ -1,0 +1,3 @@
+import { preProcessHumanizer } from './preProcessModule'
+
+export default preProcessHumanizer
