+++ src/libs/humanizer/modules/Privileges/index.ts
@@ -1,0 +1,3 @@
+import { privilegeHumanizer } from './privileges'
+
+export default privilegeHumanizer
