+++ src/libs/humanizer/modules/Sushiswap/index.ts
@@ -1,0 +1,3 @@
+import { sushiSwapModule } from './sushiSwapModule'
+
+export default sushiSwapModule
