+++ src/libs/humanizer/modules/Sushiswap/sushiSwapModule.ts
@@ -1,9 +1,9 @@
 import { Interface, ZeroAddress } from 'ethers'
 
-import { AccountOp } from '../../accountOp/accountOp'
-import { RouteProcessor } from '../const/abis'
-import { HumanizerCallModule, IrCall } from '../interfaces'
-import { getAction, getLabel, getRecipientText, getToken } from '../utils'
+import { AccountOp } from '../../../accountOp/accountOp'
+import { RouteProcessor } from '../../const/abis'
+import { HumanizerCallModule, IrCall } from '../../interfaces'
+import { getAction, getLabel, getRecipientText, getToken } from '../../utils'
 
 export const sushiSwapModule: HumanizerCallModule = (accountOp: AccountOp, irCalls: IrCall[]) => {
   const routeProcessorIface = new Interface(RouteProcessor)
