+++ src/libs/humanizer/modules/Tokens/index.ts
@@ -1,0 +1,3 @@
+import { genericErc20Humanizer, genericErc721Humanizer } from './tokens'
+
+export { genericErc20Humanizer, genericErc721Humanizer }
