+++ src/libs/humanizer/modules/Uniswap/interfaces.ts
@@ -1,0 +1,4 @@
+import { AccountOp } from '../../../accountOp/accountOp'
+import { IrCall } from '../../interfaces'
+
+export type HumanizerUniMatcher = { [key: string]: (a: AccountOp, c: IrCall) => IrCall[] }
