+++ src/libs/humanizer/modules/Uniswap/uniswap.test.ts
@@ -1,0 +1,368 @@
+import { expect } from '@jest/globals'
+
+import humanizerInfo from '../../../../consts/humanizer/humanizerInfo.json'
+import { ErrorRef } from '../../../../controllers/eventEmitter/eventEmitter'
+import { AccountOp } from '../../../accountOp/accountOp'
+import { HumanizerMeta, HumanizerVisualization, IrCall } from '../../interfaces'
+import { uniswapHumanizer } from '.'
+
+const transactions = {
+  firstBatch: [
+    // Swap exact WALLET for at least x  USDC
+    {
+      to: '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45',
+      value: BigInt(0),
+      data: '0x5ae401dc0000000000000000000000000000000000000000000000000000000064c236530000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000124b858183f00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000080000000000000000000000000b674f3fd5f43464db0448a57529eaf37f04ccea500000000000000000000000000000000000000000000003635c9adc5dea000000000000000000000000000000000000000000000000000000000000000835074000000000000000000000000000000000000000000000000000000000000004288800092ff476844f74dc2fc427974bbee2794ae002710c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f4a0b86991c6218b36c1d19d4a2e9eb0ce3606eb4800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000'
+    },
+    // Swap up to x Adex to exact DAI
+    {
+      to: '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45',
+      value: BigInt(0),
+      data: '0x5ae401dc0000000000000000000000000000000000000000000000000000000064c233bf000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000012409b8134600000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000080000000000000000000000000b674f3fd5f43464db0448a57529eaf37f04ccea50000000000000000000000000000000000000000000000056bc75e2d63100000000000000000000000000000000000000000000000000025faff1f58be30f6ec00000000000000000000000000000000000000000000000000000000000000426b175474e89094c44da98b954eedeac495271d0f000064dac17f958d2ee523a2206206994597c13d831ec7000bb8ade00c28244d5ce17d72e40330b1c318cd12b7c300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000'
+    },
+    // multicall
+    {
+      to: '0xE592427A0AEce92De3Edee1F18E0157C05861564',
+      data: '0xac9650d800000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001a00000000000000000000000000000000000000000000000000000000000000124f28c0498000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000005a5be6b067d6b5b018adbcd27ee6972105b3b4000000000000000000000000000000000000000000000000000000000064d4f15700000000000000000000000000000000000000000000048a19ce0269c802800000000000000000000000000000000000000000000000000019952df3ca0a9588000000000000000000000000000000000000000000000000000000000000002b046eee2cc3188071c02bfc1745a6b17c656e3f3d000bb8c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000412210e8a00000000000000000000000000000000000000000000000000000000',
+      value: BigInt(0)
+    }
+  ],
+  secondBatch: [
+    // first part of this has recipient 0x000...000. idk why
+    // muticall
+    {
+      to: '0xe592427a0aece92de3edee1f18e0157c05861564',
+      value: 0n,
+      data: '0xac9650d800000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000000000000000000000000000000104414bf3890000000000000000000000008a3c710e41cd95799c535f22dbae371d7c858651000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000064e5d5e7000000000000000000000000000000000000000000000ac44eff60b2f4be486300000000000000000000000000000000000000000000000001ea0706751c1289000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004449404b7c00000000000000000000000000000000000000000000000001ea0706751c1289000000000000000000000000bc5a0707cc6c731debea1f0388a4240df93259e400000000000000000000000000000000000000000000000000000000'
+    },
+    // exactInputSingle
+    {
+      to: '0xe592427a0aece92de3edee1f18e0157c05861564',
+      value: 0n,
+      data: '0x414bf389000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20000000000000000000000006e975115250b05c828ecb8ededb091975fc20a5d00000000000000000000000000000000000000000000000000000000000001f4000000000000000000000000bb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a80000000000000000000000000000000000000000000000000000000064e5ddc90000000000000000000000000000000000000000000000000d0f1a83ada48000000000000000000000000000000000000000000000000117a7744519162631a50000000000000000000000000000000000000000000000000000000000000000'
+    },
+    // exactInput
+    {
+      to: '0xe592427a0aece92de3edee1f18e0157c05861564',
+      value: 0n,
+      data: '0xc04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ca124b356bf11dc153b886ecb4596b5cb9395c410000000000000000000000000000000000000000000000000000000064e5ccb7000000000000000000000000000000000000000000000016eb3088b55b95bfa400000000000000000000000000000000000000000000000000000000925323360000000000000000000000000000000000000000000000000000000000000042ebb82c932759b515b2efc1cfbb6bf2f6dbace404002710c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f4a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000000000000000000000000000000000000000'
+    },
+    // exactOutputSingle
+    {
+      to: '0xe592427a0aece92de3edee1f18e0157c05861564',
+      value: 0n,
+      data: '0xdb3e21980000000000000000000000006e975115250b05c828ecb8ededb091975fc20a5d000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000000000000000000000000000000000000000001f4000000000000000000000000bb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a80000000000000000000000000000000000000000000000000000000064e5d7910000000000000000000000000000000000000000000000000d0f1a83ada4800000000000000000000000000000000000000000000000010594c5cd1e563664110000000000000000000000000000000000000000000000000000000000000000'
+    },
+    // problematic Call execute(bytes,bytes[],uint256) from 0xeC8...3e4
+    {
+      to: '0xeC8B0F7Ffe3ae75d7FfAb09429e3675bb63503e4',
+      value: 100000000000000n,
+      data: '0x3593564c000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000006544eb3300000000000000000000000000000000000000000000000000000000000000040b000604000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000002800000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000005af3107a40000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000005af3107a4000000000000000000000000000000000000000000000000000000000000002b7d300000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002b42000000000000000000000000000000000000060001f40b2c639c533813f4aa9d7837caf62653d097ff8500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000b2c639c533813f4aa9d7837caf62653d097ff85000000000000000000000000d4ce1f1b8640c1988360a6729d9a73c85a0c80a3000000000000000000000000000000000000000000000000000000000000000f00000000000000000000000000000000000000000000000000000000000000600000000000000000000000000b2c639c533813f4aa9d7837caf62653d097ff850000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000002b7d3'
+    },
+    {
+      to: '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45',
+      value: 0n,
+      data: '0x5ae401dc0000000000000000000000000000000000000000000000000000000065577b450000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000124b858183f0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000008000000000000000000000000002a3109c4ce8354ee771feac419b5da04ef157610000000000000000000000000000000000000000000000000000000005f5e10000000000000000000000000000000000000000000000000000ffcee5c1d6202f0000000000000000000000000000000000000000000000000000000000000042ff970a61a04b1ca14834a43f5de4533ebddb5cc80001f482af49447d8a07e3bd95bd0d56f35241523fbab10027102e9a6df78e42a30712c10a9dc4b1c8656f8f287900000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000'
+    },
+    // multicall with typed data signature
+    {
+      to: '0x643770E279d5D0733F21d6DC03A8efbABf3255B4',
+      value: 0n,
+      data: '0x3593564c000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000065cfa09900000000000000000000000000000000000000000000000000000000000000030a000c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001e0000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000001600000000000000000000000003c499c542cef5e3811e1192ce70d8cc03d5c3359000000000000000000000000ffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000000000000000000000000000065f72aff0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000643770e279d5d0733f21d6dc03a8efbabf3255b40000000000000000000000000000000000000000000000000000000065cfa50700000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000004187f8e3fd141a1d34658366ac8b3a1254c581fa087f43b539933a32561fb6638956c28fc6681ce67fa0fd946a3de61ff11633ea07572b12b1534b239e81c5600e1b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000027100000000000000000000000000000000000000000000000000024ba5664e2f5f600000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002b3c499c542cef5e3811e1192ce70d8cc03d5c33590001f40d500b1d8e8ef31e21c99d1db9a6444d3adf1270000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000024ba5664e2f5f6'
+    }
+  ]
+}
+
+const accountOp: AccountOp = {
+  accountAddr: '0xB674F3fd5F43464dB0448a57529eAF37F04cceA5',
+  networkId: 'ethereum',
+  // this may not be defined, in case the user has not picked a key yet
+  signingKeyAddr: null,
+  signingKeyType: null,
+  // this may not be set in case we haven't set it yet
+  nonce: null,
+  calls: [],
+  gasLimit: null,
+  signature: null,
+  gasFeePayment: null,
+  // This is used when we have an account recovery to finalize before executing the AccountOp,
+  // And we set this to the recovery finalization AccountOp; could be used in other scenarios too in the future,
+  // for example account migration (from v1 QuickAcc to v2)
+  accountOpToExecuteBefore: null
+  // This is fed into the humanizer to help visualize the accountOp
+  // This can contain info like the value of specific share tokens at the time of signing,
+  // or any other data that needs to otherwise be retrieved in an async manner and/or needs to be
+  // "remembered" at the time of signing in order to visualize history properly
+  // humanizerMeta: {}
+}
+
+const emitedErrors: ErrorRef[] = []
+const moockEmitError = (e: ErrorRef) => emitedErrors.push(e)
+
+describe('uniswap', () => {
+  test('uniV3', () => {
+    const expectedhumanization = [
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0x8a3c710e41cd95799c535f22dbae371d7c858651',
+          amount: 50844919041919270406243n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+          amount: 137930462904193673n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0x0000000000000000000000000000000000000000'
+        },
+        { type: 'deadline', amount: 1692784103000n }
+      ],
+      [
+        { type: 'action', content: 'Unwrap' },
+        {
+          type: 'token',
+          address: '0x0000000000000000000000000000000000000000',
+          amount: 137930462904193673n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0xbc5a0707cc6c731debea1f0388a4240df93259e4'
+        }
+      ],
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+          amount: 941000000000000000n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0x6e975115250b05c828ecb8ededb091975fc20a5d',
+          amount: 5158707941840645403045n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0xbb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a8'
+        },
+        { type: 'deadline', amount: 1692786121000n }
+      ],
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0xebb82c932759b515b2efc1cfbb6bf2f6dbace404',
+          amount: 422775565331912310692n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+          amount: 2454922038n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0xca124b356bf11dc153b886ecb4596b5cb9395c41'
+        },
+        { type: 'deadline', amount: 1692781751000n }
+      ],
+      [
+        { type: 'action', content: 'Swap up to' },
+        {
+          type: 'token',
+          address: '0x6e975115250b05c828ecb8ededb091975fc20a5d',
+          amount: 4825320403256397423633n
+        },
+        { type: 'label', content: 'for' },
+        {
+          type: 'token',
+          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+          amount: 941000000000000000n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0xbb6c8c037b9cc3bf1a4c4188d92e5d86bfce76a8'
+        },
+        { type: 'deadline', amount: 1692784529000n }
+      ],
+      [
+        { type: 'action', content: 'Wrap' },
+        {
+          type: 'token',
+          address: '0x0000000000000000000000000000000000000000',
+          amount: 100000000000000n
+        }
+      ],
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0x4200000000000000000000000000000000000006',
+          amount: 100000000000000n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0x0b2c639c533813f4aa9d7837caf62653d097ff85',
+          amount: 178131n
+        },
+        { type: 'deadline', amount: 1699015475000n }
+      ],
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
+          amount: 100000000n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0x2e9a6df78e42a30712c10a9dc4b1c8656f8f2879',
+          amount: 72003605256085551n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0x02a3109c4ce8354ee771feac419b5da04ef15761'
+        },
+        { type: 'deadline', amount: 1700232005000n }
+      ],
+      [
+        {
+          type: 'action',
+          content: 'Grant approval'
+        },
+        {
+          type: 'label',
+          content: 'for'
+        },
+        {
+          type: 'token',
+          address: '0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
+          amount: 1461501637330902918203684832716283019655932542975n
+        },
+        {
+          type: 'label',
+          content: 'to'
+        },
+
+        {
+          type: 'address',
+          address: '0x643770e279d5d0733f21d6dc03a8efbabf3255b4'
+        }
+      ],
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
+          amount: 10000n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270',
+          amount: 10337979384133110n
+        },
+        { type: 'deadline', amount: 1708105881000n }
+      ],
+      [
+        { type: 'action', content: 'Unwrap' },
+        {
+          type: 'token',
+          address: '0x0000000000000000000000000000000000000000',
+          amount: 10337979384133110n
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0x0000000000000000000000000000000000000001'
+        }
+      ]
+    ]
+    accountOp.calls = [...transactions.secondBatch]
+    let irCalls: IrCall[] = accountOp.calls
+    ;[irCalls] = uniswapHumanizer(accountOp, irCalls, humanizerInfo as HumanizerMeta, {
+      emitedError: moockEmitError
+    })
+    expect(irCalls.length).toBe(expectedhumanization.length)
+    irCalls.forEach((c, i) => {
+      expect(c?.fullVisualization?.length).toEqual(expectedhumanization[i].length)
+      c?.fullVisualization?.forEach((v: HumanizerVisualization, j: number) => {
+        expect(v).toMatchObject(expectedhumanization[i][j])
+      })
+    })
+  })
+
+  test('uniSwap', () => {
+    accountOp.calls = [...transactions.firstBatch]
+    const irCalls: IrCall[] = accountOp.calls
+    const [calls] = uniswapHumanizer(accountOp, irCalls, humanizerInfo as HumanizerMeta)
+    const expectedVisualization = [
+      [
+        { type: 'action', content: 'Swap' },
+        {
+          type: 'token',
+          address: '0x88800092ff476844f74dc2fc427974bbee2794ae',
+          amount: 1000000000000000000000n
+        },
+        { type: 'label', content: 'for at least' },
+        {
+          type: 'token',
+          address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48'
+        },
+        { type: 'deadline', amount: 1690449491000n }
+      ],
+      [
+        { type: 'action', content: 'Swap up to' },
+        {
+          type: 'token',
+          address: '0xade00c28244d5ce17d72e40330b1c318cd12b7c3'
+        },
+        { type: 'label', content: 'for' },
+        {
+          type: 'token',
+          address: '0x6b175474e89094c44da98b954eedeac495271d0f'
+        },
+        { type: 'deadline', amount: 1690448831000n }
+      ],
+      [
+        { type: 'action', content: 'Swap up to' },
+        {
+          type: 'token',
+          address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2'
+        },
+        { type: 'label', content: 'for' },
+        {
+          type: 'token',
+          address: '0x046eee2cc3188071c02bfc1745a6b17c656e3f3d'
+        },
+        { type: 'label', content: 'and send it to' },
+        {
+          type: 'address',
+          address: '0x5a5be6b067d6b5b018adbcd27ee6972105b3b400'
+        },
+        { type: 'deadline', amount: 1691677015000n }
+      ],
+      [{ type: 'action', content: 'Refund' }]
+    ]
+    expect(calls.length).toEqual(expectedVisualization.length)
+    calls.forEach((c, i) => {
+      expect(c?.fullVisualization?.length).toBe(expectedVisualization[i].length)
+      c?.fullVisualization?.forEach((v: HumanizerVisualization, j: number) => {
+        expect(v).toMatchObject(expectedVisualization[i][j])
+      })
+    })
+  })
+})
