+++ src/libs/humanizer/modules/Wrapping/index.ts
@@ -1,0 +1,3 @@
+import { wrappingModule } from './wrapping'
+
+export default wrappingModule
